Rename tool + Reorganize architecture + Introduce dev mode
# Let me introduce you : PigmentStore !

This is not final yet. However, I like the idea : each a component is a color. One only have to visit the Store in order to start a painting of their own.

# Architecture

The idea is that it such a styleguide could be for any language/library and it's not react specific.

- src/core is responsible of styleguide generation
- src/cli contains the file that will be accessible once the package is installed on other projects
- src/react contains how pigment store should behave for React components. Be careful here, it's not settled yet. Ideally, there should only be the "renderer" and "usage" which are specific to react. The rest should go back to core.

I still need to figure out how to make definitions extensible and easily related to a specific renderer.

It could even be separated into several packages. But I still too early to have such concerns.

# Dev mode

The dev mode allows to watch for changes and update the styleguide accordingly.

____________

Sorry for the one-for-all commit