Add category base on directory structure
Pretty big changes here even if it's one unique feature.

The suites are now organized by directory. Each directory matches a
Category. The name of the directory is the name that will be displayed
as the name of the Category.

This system will allow pretty cool features such as descriptions
(README.md files ?) on index pages of each Category. Moreover, it
means that it is easy to adapt the styleguide to whatever workflow
your team have (atomic design ?)