Use an actual iframe rather than an iframe defined in JS
The iframe used before was defined in the iframe container. It now
fetches an actual URL in order to stop worrying about webpack loaders
including assets in the correct file.

This move allowed me to split the core of the styleguide into... the core.
We could imagine creating a new renderer under `src/vuejs` for instance.

Browserify has no more CSS for now. I will have to build another styleguide
with its own assets in order not to make stupid mistakes.

PS: This all-in-one commit sucks :( I'll be more careful in the future