Get rid of build steps for internal node stuff
The only reason why a compile step was needed was for ES2015 `import`s.
Switched back to the good ol' `require` to avoid useless complexity.

As this tool supplies a compiled version of the styleguide, node version
shouldn't matter that much for end user. Thus I'm committed to support from
node 6.