Change how the babel plugins adds meta data
It was made on each import. However, in order to make it easier it is now
done by replacing the content of the file. This makes the final bundle
smaller since app.js won't contain the actual files of your app
and iframe wont contain the actual files of the styleguide.